Corrections to load scripts.
At least two of the three load shell scripts were incomplete. I have
attampted to clean these up slightly, although I've not tested this
yet as it was re-typed from the safe haven. Please give them a test.

I have updated the README to reflect the changes - notable, it is
expected that the config file will be sourced prior to running the
load script.